Revisiting our User model
The is_authenticated method has a misleading name. In general this
method should just return True unless the object represents a user that
should not be allowed to authenticate for some reason.

The is_active method should return True for users unless they are
inactive, for example because they have been banned.

The is_anonymous method should return True only for fake users that are
not supposed to log in to the system.

Finally, the get_id method should return a unique identifier for the
user, in unicode format. We use the unique id generated by the database
layer for this.